+++ src/tsconfig.strictNullChecks.json
@@ -213,7 +213,6 @@
 		"./vs/platform/actions/test/common/menuService.test.ts",
 		"./vs/platform/backup/common/backup.ts",
 		"./vs/platform/backup/electron-main/backupMainService.ts",
-		"./vs/platform/broadcast/electron-browser/broadcastService.ts",
 		"./vs/platform/clipboard/common/clipboardService.ts",
 		"./vs/platform/clipboard/electron-browser/clipboardService.ts",
 		"./vs/platform/commands/common/commands.ts",
@@ -666,6 +665,7 @@
 		"./vs/workbench/services/activity/common/activity.ts",
 		"./vs/workbench/services/backup/common/backup.ts",
 		"./vs/workbench/services/backup/node/backupFileService.ts",
+		"./vs/workbench/services/broadcast/electron-browser/broadcastService.ts",
 		"./vs/workbench/services/bulkEdit/electron-browser/bulkEditService.ts",
 		"./vs/workbench/services/configuration/common/configuration.ts",
 		"./vs/workbench/services/configuration/common/configurationModels.ts",
